KAN-107 refactor: products react-query hook mutation과 query로 분리
