+++ src/app/services/products/products.ts
@@ -23,25 +23,23 @@ export const postProducts = async () => {
 
 // 작품 목록 조회
 export const getProductList = async () => {
-  try {
-    const res = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/products`, {
-      method: 'GET',
-      headers: {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${process.env.NEXT_PUBLIC_ACCESS_TOKEN}`,
-      },
-      // MEMO(Sohyun): Next.js v15에서 cache의 default option은 auto no cache
-      // no cache: 개발에서는 항상 새로 fetch, 프로덕션에서는 SSG이면 빌드 시 fetch, 동적 요청이면 매 요청 fetch
-      // 작품 목록 조회의 경우 자주 변경되는 데이터가 아니므로 즉, 캐시를 저장하고 사용할때 검증(조건에 따라 fetch)하는 방식인 no cache 옵션 사용
-    })
+  const res = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/products`, {
+    method: 'GET',
+    headers: {
+      'Content-Type': 'application/json',
+      Authorization: `Bearer ${process.env.NEXT_PUBLIC_ACCESS_TOKEN}`,
+    },
+    // MEMO(Sohyun): Next.js v15에서 cache의 default option은 auto no cache
+    // no cache: 개발에서는 항상 새로 fetch, 프로덕션에서는 SSG이면 빌드 시 fetch, 동적 요청이면 매 요청 fetch
+    // 작품 목록 조회의 경우 자주 변경되는 데이터가 아니므로 즉, 캐시를 저장하고 사용할때 검증(조건에 따라 fetch)하는 방식인 no cache 옵션 사용
+  })
 
-    const data: ProductListResponseType = await res.json()
+  const data: ProductListResponseType = await res.json()
 
-    if (data.code === 'RESULT-001') {
-      return data.result
-    }
-  } catch (error) {
-    console.error('작품 목록 조회 실패: ', error)
+  if (data.code === 'RESULT-001') {
+    return data.result
+  } else {
+    throw new Error('작품 목록 조회 실패') // React-Query에서 에러를 받도록 처리
   }
 }
